feat: Integrate Cow and Fence classes into FSM
- Refactor fsmTask.cpp to use Cow& and Fence& instead of fsmTaskParams
- Update all FSM functions to accept references (runStartupRoutineFSM,
  runNormalOperationFSM, runInitializeFSM, runGreenZoneFSM,
  runStimulusZoneFSM, runFenceTransitionFSM)
- Update helper functions: dequeuedMessage, updatePosition, sendPosition,
  receivedFence, updateFence, isInFence, updateDistAndZone,
  updateAcceleration, updateState
- Add static DeviceUID deviceUID for cow identification
- Make classifyMotion() static internal function
- Update fsmTask.h with new signatures and forward declarations
- Remove duplicate type definitions (zone_t, CowState, Acceleration)
- Implement Fence class without std::vector (static arrays)
- Add Cow and Fence test suites
- Remove old documentation files from root (moved to docs/)

+++ CM4/Core/Src/cow.cpp
@@ -25,7 +25,7 @@ void Cow::updateCurrentZone(zone_t zone) {
     currentZone = zone;
 }
 
-void Cow::updateDistanceToLimit(float distance) {
+void Cow::updateDistanceToLimit(double distance) {
     distanceToLimit = distance;
 }
 
@@ -49,7 +49,6 @@ zone_t Cow::getCurrentZone() const {
     return currentZone;
 }
 
-float Cow::getDistanceToLimit() const
-{
+double Cow::getDistanceToLimit() const {
     return distanceToLimit;
 }
